+++ changelog.md
@@ -32,6 +32,7 @@ Codion Change Log
 - EntityEditComponentPanel.createLabel() added for read-only value linking.
 - EntityEditComponentPanel.InputPanelBuilder.label() and component() overloaded with Supplier parameter.
 - EntityEditComponentPanel.InputPanelBuilder now extends ComponentBuilder.
+- EntityEditComponentPanel.InputPanelBuilder.label() overloads with constraints removed.
 
 ## 0.18.43
 ### is.codion.common.core